% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hru_pw.R
\name{add_start_end}
\alias{add_start_end}
\title{Add start and end of a crop planting period at beginning and end of year for
winter crops. This is required to correctly display crops in plot_hru_pw_day}
\usage{
add_start_end(tbl)
}
\arguments{
\item{tbl}{Table with crops and plant and harvest dates}
}
\description{
Add start and end of a crop planting period at beginning and end of year for
winter crops. This is required to correctly display crops in plot_hru_pw_day
}
\keyword{internal}
