% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_soft_cal.R
\name{build_wb_softcal_input}
\alias{build_wb_softcal_input}
\title{Build/load the water balance soft calibration input files and write them
into the folder where the model is executed.}
\usage{
build_wb_softcal_input(run_path, wateryield_ratio, baseflow_ratio)
}
\arguments{
\item{run_path}{Path to the copy of SWAT+ project where the soft calibration
will be executed}

\item{wateryield_ratio}{Numeric value between 0 and 1 which defines the
fraction of precipitation which becomes water yield through fast, slow,
and aquifer flow}

\item{baseflow_ratio}{Numeric value between 0 and 1 which defines the
fraction of water yield which is base flow}
}
\description{
Build/load the water balance soft calibration input files and write them
into the folder where the model is executed.
}
\keyword{internal}
