% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_hru_waterbalance.R
\name{check_hru_waterbalance}
\alias{check_hru_waterbalance}
\title{Check average annual water balance simulations at the HRU level}
\usage{
check_hru_waterbalance(
  sim_verify,
  check = c("precip", "et", "runoff", "sw", "cn"),
  ignore_lum = NULL,
  add_values = TRUE
)
}
\arguments{
\item{sim_verify}{Simulation output of the function \code{run_swat_verification()}.
To plot the heat units at least the output option \code{outputs = 'mgt'} must
be set in  \code{run_swat_verification()}}

\item{check}{Character vector to indicate which water balance components should
be checked. Options are 'precip' to check precipitation, 'et' to check
evapotranspiration and components of ET, 'runoff' to check runoff components
such as 'wyld' (water yield), surq (surface runoff), or 'perc' (percolation),
'sw' to check the soil water content, and 'cn' to check curve number values.}

\item{ignore_lum}{Character vector to define land uses which should be ignored
in the water balance checks.}

\item{add_values}{Boolean, to define wether to only show the checks (FALSE),
or to also include the respective average annual water balance values.}
}
\value{
Returns a table of HRUs for which at least one of the selected checks
  identified a potential issue.
}
\description{
This function applies check rules which were proposed by White et al. (2014)
and which are implemented in SWATCheck for average annual water balance
output variables. Different to SWATCheck the checks here are performed for
all or only selected HRUs
}
