% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mgt_harv.R
\name{diff1}
\alias{diff1}
\title{Concatenate the first value of a vector and the differences of following values.}
\usage{
diff1(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector
}
\description{
Concatenate the first value of a vector and the differences of following values.
}
\keyword{internal}
