% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_soft_cal.R
\name{download_sft_files}
\alias{download_sft_files}
\title{downloads the required .sft files from some source
printing just for diagnostics, maybe remove.}
\usage{
download_sft_files(path)
}
\arguments{
\item{path}{text string to (temporary) directory}
}
\description{
downloads the required .sft files from some source
printing just for diagnostics, maybe remove.
}
\keyword{internal}
