% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mgt_harv.R
\name{filter_hru_at_harvkill}
\alias{filter_hru_at_harvkill}
\title{Filter HRUs by their variable values at harvest/kill}
\usage{
filter_hru_at_harvkill(sim_verify, ...)
}
\arguments{
\item{sim_verify}{Simulation output of the function \code{run_swat_verification()}
as a single object or as a list of objects (representing \code{run_swat_verification()}
function runs with different settings).
To plot the heat units at least the output option \code{outputs = 'mgt'}
must be set in \code{run_swat_verification()}.}

\item{...}{Boolean operations to be applied in the filter operation of the
management simulation outputs. Possible variables are \code{crop, phu, plant_bioms,
  yield, water_stress, aero_stress, temp_stress, n_stress} and \code{p_stress}.}
}
\description{
The function can be complementary to \code{plot_variable_at_harvkill()} plots.
If any issues are identified in the boxplots, the results can be filtered for
further analysis with \code{filter_hru_at_harvkill()}.
}
\examples{
\dontrun{
# Filter all HRUs with unusually high PHU values at harvest kill
filter_hru_at_harvkill(sim_verify, phu > 5)

# Filter all HRUs with crop 'wbar' or 'wira' planted but no yield
filter_hru_at_harvkill(sim_verify, crop \%in\% c('wbar', 'wira'), yield == 0)
}
}
