% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_landuse_labels.R
\name{find_no_mgt}
\alias{find_no_mgt}
\title{Find the managments with 0 management operations}
\usage{
find_no_mgt(mgt)
}
\arguments{
\item{mgt}{Management schedule input table in tibble format}
}
\description{
Find the managments with 0 management operations
}
\keyword{internal}
