% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_swat_exe}
\alias{find_swat_exe}
\title{Find the SWAT+ executable file and trigger error if 0 or more than 1
file are found.}
\usage{
find_swat_exe(project_path, os)
}
\arguments{
\item{project_path}{Path to the SWAT+ project folder (i.e. TxtInOut).}

\item{os}{Text string that defines the operating system}
}
\description{
Find the SWAT+ executable file and trigger error if 0 or more than 1
file are found.
}
\keyword{internal}
