% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_file_version}
\alias{get_file_version}
\title{Get the Highest Version Number of a File}
\usage{
get_file_version(parent_dir, file_name, add)
}
\arguments{
\item{parent_dir}{A string representing the path to the directory where the
files are located.}

\item{file_name}{A string representing the setup name.}

\item{add}{A string to be added to the file name pattern to match.}
}
\value{
The highest version number found among the matching files as a
numeric value. Returns 0 if no matching files are found.
}
\description{
This function searches for files in a specified directory that match a given
pattern and extracts the highest version number from those files.
}
\examples{
\dontrun{
file_version <- get_file_version(parent_dir, save_name, '_nostr_') + 1
}
}
