% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hru_pw.R
\name{get_hru_id_by_attribute}
\alias{get_hru_id_by_attribute}
\title{Get HRU ids by filtering for HRU attributes}
\usage{
get_hru_id_by_attribute(sim_verify, lum = NULL, mgt = NULL, soil = NULL)
}
\arguments{
\item{sim_verify}{Simulation output of the function \code{run_swat_verification()}.
To plot the heat units at least the output option \code{outputs = 'mgt'} must
be set in  \code{run_swat_verification()}}

\item{lum}{Optional character vector with landuse.lum labels}

\item{mgt}{Optional character vector with management labels as defined in management.sch.}

\item{soil}{Optional character vector with soil type labels as defined in the soil data.}
}
\value{
Returns a table with hru ids and attributes.
}
\description{
get_hru_id_by_attribute Uses the hru-data and landuse_lum saved after running
\code{run_swat_verification()} and filters the attributes by the defined
filter attributes.
}
