% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_soft_cal.R
\name{modify_wb_parms}
\alias{modify_wb_parms}
\title{Modify ‘water_balance.sft’ with values for fractions. Only the values WYR and
BFR columns need to be modified}
\usage{
modify_wb_parms(path)
}
\arguments{
\item{path}{string to (temporary) directory}
}
\description{
Modify ‘water_balance.sft’ with values for fractions. Only the values WYR and
BFR columns need to be modified
}
\keyword{internal}
