% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_climate.R
\name{plot_basin_var}
\alias{plot_basin_var}
\title{Plot selected basin variable aggregated by defined time step}
\usage{
plot_basin_var(sim_verify, var, period = "day", fn_summarize = "mean")
}
\arguments{
\item{sim_verify}{Simulation output of the function \code{run_swat_verification()}.
To plot the climate outputs at least the output option \code{outputs = 'wb'} must
be set in  \code{run_swat_verification()}.}

\item{var}{Character which defines the variable to be printed}

\item{period}{(optional) character describing, which time interval to display (default is "day",
other examples are "week", "month", etc). \code{Default = "day"}}

\item{fn_summarize}{(optional) function to recalculate to time interval (default is "mean", other examples
are "median", "sum", etc). \code{Default = "mean"}}
}
\value{
plotly figure object
}
\description{
Plot selected basin variable aggregated by defined time step
}
\examples{
\dontrun{
plot_basin_var(sim_nostress, "percn")
}
}
