% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_climate.R
\name{plot_climate_annual}
\alias{plot_climate_annual}
\title{Plot annual overview of simulated climate variables of a SWAT+ simulation}
\usage{
plot_climate_annual(sim_verify)
}
\arguments{
\item{sim_verify}{Simulation output of the function \code{run_swat_verification()}.
To plot the climate outputs at least the output option \code{outputs = 'wb'} must
be set in  \code{run_swat_verification()}.}
}
\value{
Returns a patchwork of ggplots for the annual values of the simulated
  climate variables.
}
\description{
plot_climate_annual uses the simulated basin averages of pet, the different
et fractions, precip, snofall, temperatures, and solar radiation to plot
annual sums and averages.
The aim of the function is to provide an overview of the simulated variables
and help to identify potential issues with the model weather input data.
}
