% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hru_pw.R
\name{plot_hru_pw}
\alias{plot_hru_pw}
\title{Aggregate and plot simulated variables saved in hru_pw_day}
\usage{
plot_hru_pw(
  sim_verify,
  hru_id,
  var,
  title = "",
  period = "day",
  fn_summarize = "mean",
  interactive = FALSE
)
}
\arguments{
\item{sim_verify}{Simulation output of the function \code{run_swat_verification()}.
To plot the heat units at least the output option \code{outputs = 'mgt'} must
be set in  \code{run_swat_verification()}.}

\item{hru_id}{Numeric vector with HRU ids for which variables should be plotted.}

\item{var}{Character vector that defines the variable names that are plotted.}

\item{title}{Character for title to be put in the figure.}

\item{period}{(optional) character describing, which time interval to display (default is "day",
other examples are "week", "month", etc). \code{Default = "day"}}

\item{fn_summarize}{(optional) function to recalculate to time interval (default is "mean", other examples
are "median", "sum", etc). \code{Default = "mean"}}

\item{interactive}{(optional) Boolean TRUE to provide output as plotly object, FALSE - as ggplot.  \code{Default = "FALSE"}}
}
\value{
ggplot or plotly object
}
\description{
Aggregate and plot simulated variables saved in hru_pw_day
}
\examples{
\dontrun{
id <- get_hru_id_by_attribute(sim_nostress, "wwht_lum")
plot_hru_pw(sim_nostress, sample(id$id, 10), "lai")
plot_hru_pw(sim_nostress, hru_id = c(2, 10), var = c("lai", "bioms"),
title = "My great figure", period = "year", fn_summarize = "mean", interactive = TRUE)
}
}
