% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hru_pw.R
\name{plot_hru_pw_day}
\alias{plot_hru_pw_day}
\title{Plot daily simulated variables which are saved in hru_pw_day}
\usage{
plot_hru_pw_day(sim_verify, hru_id, var, years = 1900:2100, add_crop = TRUE)
}
\arguments{
\item{sim_verify}{Simulation output of the function \code{run_swat_verification()}.
To plot the heat units at least the output option \code{outputs = 'mgt'} must
be set in  \code{run_swat_verification()}}

\item{hru_id}{Numeric vector with HRU ids for which variables should be plotted}

\item{var}{Character vector that defines the variable names that are plotted}

\item{years}{Years of the simulated data for which varaibles are plotted.}

\item{add_crop}{TRUE adds subplot with growing periods of crops in plotted HRUs.}
}
\value{
Returns a ggplot or patchwork ggplot of daily hru_pw output variables
}
\description{
plot_hru_pw_day plots daily time series of variables from the output file hru_pw_day
}
\examples{
\dontrun{
plot_hru_pw_day(sim_nostress, hru_id = 1, var = c('lai', 'bioms'), add_crop = TRUE)
}

}
