% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hru_pw.R
\name{plot_hru_var}
\alias{plot_hru_var}
\title{Aggregate and plot simulated variables saved in hru_pw_day}
\usage{
plot_hru_var(sim_verify, hru_id, var, period = "day", fn_summarize = "mean")
}
\arguments{
\item{sim_verify}{Simulation output of the function \code{run_swat_verification()}.
To plot the heat units at least the output option \code{outputs = 'mgt'} must
be set in  \code{run_swat_verification()}.}

\item{hru_id}{Numeric vector with HRU ids for which variables should be plotted.}

\item{var}{Character vector that defines the variable names that are plotted.}

\item{period}{(optional) character describing, which time interval to display (default is "day",
other examples are "week", "month", etc). \code{Default = "day"}}

\item{fn_summarize}{(optional) function to recalculate to time interval (default is "mean", other examples
are "median", "sum", etc). \code{Default = "mean"}}
}
\value{
plotly figure object
}
\description{
Aggregate and plot simulated variables saved in hru_pw_day
}
\examples{
\dontrun{
id <- get_hru_id_by_attribute(sim_nostress, "wwht_lum")
plot_hru_var(sim_nostress, sample(id$id, 10), "lai")
}
}
