% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hru_pw.R
\name{plot_hru_var_aa}
\alias{plot_hru_var_aa}
\title{Plot simulated variables of filtered HRUs saved in hru_wb_aa with boxplots}
\usage{
plot_hru_var_aa(sim_verify, lum = NULL, mgt = NULL, soil = NULL)
}
\arguments{
\item{sim_verify}{Simulation output of the function \code{run_swat_verification()}.
To plot the heat units at least the output option \code{outputs = 'wb'} must
be set in  \code{run_swat_verification()}}

\item{lum}{Optional character vector with landuse.lum labels}

\item{mgt}{Optional character vector with management labels as defined in management.sch.}

\item{soil}{Optional character vector with soil type labels as defined in the soil data.}
}
\value{
plot_ly multi boxplot figure for all available variables, which are not 0.
}
\description{
Plot simulated variables of filtered HRUs saved in hru_wb_aa with boxplots
}
\examples{
\dontrun{
plot_hru_var_aa(sim_nostress, "wwht_lum")
}
}
