% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_climate.R
\name{plot_monthly_snow}
\alias{plot_monthly_snow}
\title{Plot average monthly precipitation components rainfall and snofall and the snow melt}
\usage{
plot_monthly_snow(sim_verify)
}
\arguments{
\item{sim_verify}{Simulation output of the function \code{run_swat_verification()}.
To plot the climate outputs at least the output option \code{outputs = 'wb'} must
be set in  \code{run_swat_verification()}.}
}
\value{
Returns a ggplots for the monthly average values of the simulated
precip, snofall and snomlt.
}
\description{
plot_snow_monthly uses the simulated basin averages of precip, and snofall,
and plots average monthly sums of rainfall, snofall and snomlt.
This is plot can be helpful to verify the dominant snow processes and
can be useful in catchments where large spring runoff through snowmelt
plays a relevant role.
}
