% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mgt.R
\name{plot_phuplant_harvest}
\alias{plot_phuplant_harvest}
\title{Boxplot of heat unit fractions for crops at harvest-kill operation}
\usage{
plot_phuplant_harvest(sim_verify)
}
\arguments{
\item{sim_verify}{Simulation output of the function \code{run_swat_verification()}.
To plot the heat units at least the output option \code{outputs = 'mgt'} must
be set in  \code{run_swat_verification()}}
}
\value{
ggplot boxplot with crop heat unit fractions at harvest-kill.
}
\description{
plot_phuplant_harvest plots boxplots of the collected heat unit fractions
at harvest-kill of a crop separated for all identified crops.
}
