% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ps_tile.R
\name{plot_ps}
\alias{plot_ps}
\title{Plot yearly simulated point source values (as yearly loads or average concentrations)}
\usage{
plot_ps(sim_verify, conc = FALSE)
}
\arguments{
\item{sim_verify}{Simulation output of the function \code{run_swat_verification()}.
To plot the heat units at least the output option \code{outputs = 'wb'} must
be set in  \code{run_swat_verification()}}

\item{conc}{Boolean, TRUE to provide figure for point source average yearly pollutant concentrations,
FALSE - for average yearly pollutant loads.}
}
\value{
ggplot object for point source yearly simulated data
}
\description{
Plot yearly simulated point source values (as yearly loads or average concentrations)
}
\examples{
\dontrun{
plot_ps(sim_nostress, TRUE)
}
}
