% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ps_tile.R
\name{plot_qtile}
\alias{plot_qtile}
\title{Plot tile drain flow histogram and distribution curve}
\usage{
plot_qtile(
  sim_verify,
  exclude_lum = c("urhd_lum", "urmd_lum", "urml_lum", "urld_lum", "ucom_lum", "uidu_lum",
    "utrn_lum", "uins_lum", "urbn_lum")
)
}
\arguments{
\item{sim_verify}{Simulation output of the function \code{run_swat_verification()}.
To plot the heat units at least the output option \code{outputs = 'wb'} must
be set in  \code{run_swat_verification()}}

\item{exclude_lum}{Character vector to define land uses which are excluded
in the printed table.}
}
\value{
Returns ggplot object with histogram and density curve for tile drain from distribution.
}
\description{
Plot tile drain flow histogram and distribution curve
}
\examples{
\dontrun{
plot_qtile(sim_nostress)
}
}
