% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_climate.R
\name{plot_stat_text}
\alias{plot_stat_text}
\title{Plot annual average values in tabular text form as a ggplot}
\usage{
plot_stat_text(tbl, vars, pos, unit_add, add_title)
}
\arguments{
\item{tbl}{Processed climate variable table}

\item{vars}{Character vector which defines the variables to be printed}

\item{pos}{Print positions of the variables and their values}

\item{unit_add}{Character string which defines unit to be added to values}

\item{add_title}{Boolean to define if title should be added to table}
}
\description{
Plot annual average values in tabular text form as a ggplot
}
\keyword{internal}
