% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hru_pw.R
\name{plot_water_partition}
\alias{plot_water_partition}
\title{Plot simulated variables of water partition of filtered HRUs saved in hru_wb_aa}
\usage{
plot_water_partition(
  sim_verify,
  tile = NULL,
  lum = NULL,
  mgt = NULL,
  soil = NULL,
  exclude_lum = c("urhd_lum", "urmd_lum", "urml_lum", "urld_lum", "ucom_lum", "uidu_lum",
    "utrn_lum", "uins_lum", "urbn_lum"),
  boxpoints = TRUE
)
}
\arguments{
\item{sim_verify}{Simulation output of the function \code{run_swat_verification()}.
To plot the heat units at least the output option \code{outputs = 'wb'} must
be set in  \code{run_swat_verification()}}

\item{tile}{Optional Boolean TRUE for selecting HRUs with working tiles, FALSE - without working tiles and NULL for selecting all HRUs.}

\item{lum}{Optional character vector with landuse.lum labels}

\item{mgt}{Optional character vector with management labels as defined in management.sch.}

\item{soil}{Optional character vector with soil type labels as defined in the soil data.}

\item{exclude_lum}{Character vector to define land uses which are excluded
in the printed table.}

\item{boxpoints}{Optional Boolean TRUE for displaying outliers, FALSE for hiding them.
\code{Default = TRUE}}
}
\value{
plotly figure object
}
\description{
Plot simulated variables of water partition of filtered HRUs saved in hru_wb_aa
}
\examples{
\dontrun{
plot_water_partition(sim_nostress, tile = TRUE)
}
}
