% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mgt_harv.R
\name{prepare_biomass}
\alias{prepare_biomass}
\title{Prepare mgt_out data for plotting biomass}
\usage{
prepare_biomass(mgt_out, years)
}
\arguments{
\item{mgt_out}{Data frame with management output data}

\item{years}{Integer vector of years for which values should be calculated}
}
\value{
data frame with crop names and bio mass
}
\description{
Biomass is only extracted for harvest/kill operations.
}
\keyword{internal}
