% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mgt_harv.R
\name{prepare_phu}
\alias{prepare_phu}
\title{Prepare mgt_out data for plotting PHU fractions}
\usage{
prepare_phu(mgt_out, years)
}
\arguments{
\item{mgt_out}{Data frame with management output data}

\item{years}{Integer vector of years for which values should be calculated}
}
\value{
data frame with crop names and phu fractions
}
\description{
PHU fractions are extracted for the last operation before a kill operation is set.
}
\keyword{internal}
