% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mgt_harv.R
\name{prepare_stress}
\alias{prepare_stress}
\title{Prepare mgt_out data for plotting stress factors}
\usage{
prepare_stress(mgt_out, years)
}
\arguments{
\item{mgt_out}{Data frame with management output data}

\item{years}{Integer vector of years for which values should be calculated}
}
\value{
data frame with crop names, stress names and values
}
\description{
For static land uses (no plant and kill operations) annual changes in stress
values are calculated (annual increase of a stress factor).
For crops with defined growing periods the stress values for the last harvest
operation is extracted.
}
\keyword{internal}
