% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mgt_harv.R
\name{prepare_var_at_harvkill}
\alias{prepare_var_at_harvkill}
\title{Prepare dataframe for plot_variable_at_harvkill function}
\usage{
prepare_var_at_harvkill(mgt_out, years, variable)
}
\arguments{
\item{mgt_out}{Data frame with management output data}

\item{years}{Integer vector of years for which values should be calculated}

\item{variable}{String to define which variable is selected for plotting}
}
\value{
data frame for plotting
}
\description{
Prepare dataframe for plot_variable_at_harvkill function
}
\examples{
\dontrun{
var_at_harv(sim_nostress$mgt_out, 1900:2100)
}
}
\keyword{internal}
