% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mgt_harv.R
\name{prepare_yield}
\alias{prepare_yield}
\title{Prepare mgt_out data for plotting yields}
\usage{
prepare_yield(mgt_out, years)
}
\arguments{
\item{mgt_out}{Data frame with management output data}

\item{years}{Integer vector of years for which values should be calculated}
}
\value{
data frame with crop names and yields
}
\description{
For static land uses (no plant and kill operations) annual yields are calculated.
For crops with defined growing periods the yield sums of that period is calculated.
}
\keyword{internal}
