% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hru_pw.R
\name{print_avannual_qtile}
\alias{print_avannual_qtile}
\title{Print the average annual qtile for HRUs}
\usage{
print_avannual_qtile(
  sim_verify,
  exclude_lum = c("urhd_lum", "urmd_lum", "urml_lum", "urld_lum", "ucom_lum", "uidu_lum",
    "utrn_lum", "uins_lum", "urbn_lum")
)
}
\arguments{
\item{sim_verify}{Simulation output of the function \code{run_swat_verification()}.
To plot the heat units at least the output option \code{outputs = 'wb'} must
be set in  \code{run_swat_verification()}}

\item{exclude_lum}{Character vector to define land uses which are excluded
in the printed table.}
}
\value{
Returns a table with hru ids average annual qtile and attributes.
}
\description{
print_avannual_qtile prints a table with the average annual qtile in mm
for HRUs that used a tile flow parametrization in landuse.lum
}
