% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mgt.R
\name{print_triggered_mgt}
\alias{print_triggered_mgt}
\title{Print the triggered management operations from the mgt outputs for a certain HRU}
\usage{
print_triggered_mgt(sim_verify, hru_id, start_year = 1900, end_year = 2100)
}
\arguments{
\item{sim_verify}{Simulation output of the function \code{run_swat_verification()}.
To print the management at least the output option \code{outputs = 'mgt'} must
be set in  \code{run_swat_verification()}}

\item{hru_id}{id of the HRU for which the triggered management operations should
be printed}

\item{start_year}{Integer value to define the first year of printing.}

\item{end_year}{Integer value to define the last year of printing.}
}
\value{
Prints a tibble with triggered operations to the R console.
}
\description{
print_triggered_mgt extracts the triggered management operation which are written
into mgt_out.txt and reformats them to resemble the scheduled operations in
managment.sch to ease a comparison between them.
}
