% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_swat_verify.R
\name{read_mgt}
\alias{read_mgt}
\title{Read SWAT+ management output file and return the read output in a tibble}
\usage{
read_mgt(run_path)
}
\arguments{
\item{run_path}{Path to the folder where simulations are performed}
}
\description{
Read SWAT+ management output file and return the read output in a tibble
}
\keyword{internal}
