% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_landuse_labels.R
\name{read_plant_ini}
\alias{read_plant_ini}
\title{Read function for the plant.ini input file}
\usage{
read_plant_ini(path)
}
\arguments{
\item{path}{Path to the SWAT+ project folder (i.e. TxtInOut).}
}
\description{
Read function for the plant.ini input file
}
\keyword{internal}
