% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_swat_verify.R
\name{read_sch}
\alias{read_sch}
\title{Read SWAT+ management schedule file and return the read file in a tibble}
\usage{
read_sch(run_path)
}
\arguments{
\item{run_path}{Path to the folder where simulations are performed}
}
\description{
Read SWAT+ management schedule file and return the read file in a tibble
}
\keyword{internal}
