% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_swat_verify.R
\name{read_tbl}
\alias{read_tbl}
\title{Read SWAT+ output that is arranged in a tabular format (most outputs)
and return the read output table in a tibble format}
\usage{
read_tbl(out_file, run_path)
}
\arguments{
\item{out_file}{Name of the output file that should be read.}

\item{run_path}{Path to the folder where simulations are performed}
}
\description{
Read SWAT+ output that is arranged in a tabular format (most outputs)
and return the read output table in a tibble format
}
\keyword{internal}
