% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_swat_verify.R
\name{read_wb_aa}
\alias{read_wb_aa}
\title{reads and re-formats the SWAT+ soft calibration output}
\usage{
read_wb_aa(path)
}
\arguments{
\item{path}{string to (temporary) directory}
}
\value{
returns a tibble of the formatted output of the soft-cal routine
}
\description{
reads and re-formats the SWAT+ soft calibration output
}
\keyword{internal}
