% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{remove_tail}
\alias{remove_tail}
\title{Remove tail endings for management, plant community, land use classes}
\usage{
remove_tail(f, pattern)
}
\arguments{
\item{f}{multiline character}

\item{pattern}{pattern after which to remove tail}
}
\value{
corrected multiline character
}
\description{
This function is required for SWATfarmR generated management files
as management.sch, plant.ini, landuse.lum and hru-data.hru can get too
long names for swat executable.
}
\examples{
\dontrun{
library(readr)
landuse <- read_lines(paste0(project_path,'/landuse.lum'), lazy = FALSE)
landuse  <- remove_tail(landuse, "lum")
landuse  <- remove_tail(landuse, "comm")
landuse  <- remove_tail(landuse, "mgt")
write_lines(landuse, paste0(project_path,'/landuse_new.lum'))
}
}
\keyword{internal}
