% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_landuse_labels.R
\name{replace_ini_lbl}
\alias{replace_ini_lbl}
\title{Replace the plant ini label in a line of the plant.ini input file}
\usage{
replace_ini_lbl(ini_line, ini_lbl_upd)
}
\arguments{
\item{ini_line}{Text line from the plant.ini input file}

\item{ini_lbl_upd}{New plant.ini label for replacement of the old label}
}
\description{
Replace the plant ini label in a line of the plant.ini input file
}
\keyword{internal}
