% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mgt.R
\name{report_mgt}
\alias{report_mgt}
\title{Generate a report table that compares the scheduled and triggered managements}
\usage{
report_mgt(sim_verify, write_report = FALSE)
}
\arguments{
\item{sim_verify}{Simulation output of the function \code{run_swat_verification()}.
To print the management at least the output option \code{outputs = 'mgt'} must
be set in  \code{run_swat_verification()}}

\item{write_report}{(optional) Boolean TRUE for writing output to 'schedule_report.txt' file,
FALSE - not preparing this file. Default \code{write_report = FALSE}.}
}
\value{
Returns a tibble that summarises all management schedules for
  which operations where scheduled, that were either not triggered of
  for which operation properties differ..
}
\description{
report_mgt compares the scheduled management operations for all schedules
with the triggered management operations. Therefore HRUs are randomly
selected for each schedule where one of the schedules is implemented.
}
