% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_soft_cal.R
\name{run_softcal_waterbalance}
\alias{run_softcal_waterbalance}
\title{Runs the SWAT+ soft-calibration routine for water balance components.}
\usage{
run_softcal_waterbalance(
  project_path,
  wateryield_ratio,
  baseflow_ratio,
  start_date = NULL,
  end_date = NULL,
  years_skip = NULL,
  keep_folder = FALSE
)
}
\arguments{
\item{project_path}{Path to the SWAT+ project folder (i.e. TxtInOut).}

\item{wateryield_ratio}{Numeric value between 0 and 1 which defines the
fraction of precipitation which becomes water yield through fast, slow,
and aquifer flow.}

\item{baseflow_ratio}{Numeric value between 0 and 1 which defines the
fraction of water yield which is base flow.}

\item{start_date}{(optional) Start date of the SWAT simulation. Provided as
character string in any ymd format (e.g. 'yyyy-mm-dd'), numeric value
in the form yyyymmdd, or in Date format.}

\item{end_date}{(optional) End date of the SWAT simulation. Provided as
character string in any ymd format (e.g. 'yyyy-mm-dd'), numeric value
in the form yyyymmdd, or in Date format.}

\item{years_skip}{(optional) Integer value to define the number of simulation
years that are skipped before writing SWAT model outputs.}

\item{keep_folder}{(optional) If \code{keep_folder = TRUE}
'.model_run/verification' is kept and not deleted after finishing model runs.
In this case '.model_run' is reused in a new model run if \code{refresh = FALSE}.
\code{Default = FALSE}}
}
\value{
Returns the simulation results for the soft calibration routine as a
  tibble.
}
\description{
\code{run_softcal_waterbalance} executes the SWAT+ soft calibration
routine to adjust the overall water balance of a SWAT+ model setup.
The user can specify the \code{wateryield_ratio} and the
\code{baseflow_ratio} of the simulated watershed. Based on the provided
ratios the routine adjusts the respective model parameters to improve
the fit of the simulated water balance components to the provided ratios.
}
