% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_swat_verify.R
\name{run_swat_verification}
\alias{run_swat_verification}
\title{Run simulations for model verification}
\usage{
run_swat_verification(
  project_path,
  outputs = c("wb", "mgt", "plt"),
  start_date = NULL,
  end_date = NULL,
  years_skip = NULL,
  nostress = 1,
  keep_folder = FALSE
)
}
\arguments{
\item{project_path}{Path to the SWAT+ project folder (i.e. TxtInOut).}

\item{outputs}{Define the outputs that should be read after the simulation
run. The outputs that are defined here depend on the verification steps
that should be performed on the outputs.}

\item{start_date}{(optional) Start date of the SWAT simulation. Provided as
character string in any ymd format (e.g. 'yyyy-mm-dd'), numeric value
in the form yyyymmdd, or in Date format.}

\item{end_date}{(optional) End date of the SWAT simulation. Provided as
character string in any ymd format (e.g. 'yyyy-mm-dd'), numeric value
in the form yyyymmdd, or in Date format.}

\item{years_skip}{(optional) Integer value to define the number of simulation
years that are skipped before writing SWAT model outputs.}

\item{nostress}{nostress parameter in the 'codes.bsn' file to activate/deactivate
plant stresses for plant growth. Set \code{nostress = 0} to activate all stress
factors, \code{nostress = 1} to deactivate all stress factors, and \code{nostress = 2}
to only activate nutrient plant stress.}

\item{keep_folder}{(optional) If \code{keep_folder = TRUE}
'.model_run/verification' is kept and not deleted after finishing model runs.
In this case '.model_run' is reused in a new model run if \code{refresh = FALSE}.
\code{Default = FALSE}}
}
\value{
Returns the simulation results for the defined output variables as a
  list of tibbles.
}
\description{
This function allows to run a SWAT2012 project in R.
Basic settings for the SWAT run such as the simulation period or the time
interval for the outputs can be done directly. SWAT simulation outputs can be
defined that are returned in a 'tidy' format in R. Functionality such as model
parametrization, parallel execution of simulations, or incremental saving of
simulation runs is provided.
}
