% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_swat_verify.R
\name{set_codes_bsn}
\alias{set_codes_bsn}
\title{Set the nostress value in the code.bsn file}
\usage{
set_codes_bsn(run_path, nostress)
}
\arguments{
\item{run_path}{Path to the folder where simulations are performed}

\item{nostress}{nostress parameter in the 'codes.bsn' file to activate/deactivate
plant stresses for plant growth. Set \code{nostress = 1} to activate all stress
factors, \code{nostress = 1} to deactivate all stress factors, and \code{nostress = 1}
to only activate nutrient plant stress.}
}
\description{
Set the nostress value in the code.bsn file
}
\keyword{internal}
