% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_input_files.R
\name{set_time_sim}
\alias{set_time_sim}
\title{Read and set SWAT+ time.sim file write the updated time.sim and the original file
as a backup 'time_backup.sim'}
\usage{
set_time_sim(project_path, run_path, start_date, end_date)
}
\arguments{
\item{project_path}{Path to the SWAT project folder (i.e. TxtInOut)}

\item{run_path}{Path to the folder where simulations are performed}

\item{start_date}{(optional) Start date of the SWAT simulation. Provided as
character string in any ymd format (e.g. 'yyyy-mm-dd'), numeric value
in the form yyyymmdd, or in Date format.}

\item{end_date}{(optional) End date of the SWAT simulation. Provided as
character string in any ymd format (e.g. 'yyyy-mm-dd'), numeric value
in the form yyyymmdd, or in Date format.}
}
\description{
Read and set SWAT+ time.sim file write the updated time.sim and the original file
as a backup 'time_backup.sim'
}
\keyword{internal}
