% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_soft_cal.R
\name{toggle_sft}
\alias{toggle_sft}
\title{toggles the soft calibration routine either ON or OFF}
\usage{
toggle_sft(path, switch)
}
\arguments{
\item{path}{string to (temporary) directory}

\item{switch}{string, either "off" or "on"}
}
\description{
toggles the soft calibration routine either ON or OFF
}
\keyword{internal}
