% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mgt_harv.R
\name{var_at_harv}
\alias{var_at_harv}
\title{Prepare dataframe for plot_variable_at_harvkill function}
\usage{
var_at_harv(mgt_out, years)
}
\arguments{
\item{mgt_out}{dataframe with model output data}

\item{years}{integer vector to filter dataframe}
}
\value{
dataframe for plotting
}
\description{
Prepare dataframe for plot_variable_at_harvkill function
}
\examples{
\dontrun{
var_at_harv(sim_nostress$mgt_out, 1900:2100)
}
}
\keyword{internal}
