% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_landuse_labels.R
\name{write_tbl_file}
\alias{write_tbl_file}
\title{Write tabular input file in SWAT+ table text fiel}
\usage{
write_tbl_file(tbl, head, fmt, path)
}
\arguments{
\item{tbl}{Parameter input table}

\item{head}{header lines of input file}

\item{fmt}{vector of formats in which the table columns should be written}

\item{path}{Path of the input file in the project folder}
}
\description{
Write tabular input file in SWAT+ table text fiel
}
\keyword{internal}
