% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_swat_verify.R
\name{build_model_run}
\alias{build_model_run}
\title{Generate folder structure for SWAT execution}
\usage{
build_model_run(project_path, folder_name)
}
\arguments{
\item{project_path}{Path to the SWAT project folder (i.e. TxtInOut)}

\item{folder_name}{Name of the folder in which simulations are performed}
}
\description{
Generate folder structure for SWAT execution
}
\keyword{internal}
