% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_swat_verify.R
\name{build_model_run}
\alias{build_model_run}
\title{Generate folder structure for SWAT execution}
\usage{
build_model_run(project_path)
}
\arguments{
\item{project_path}{Path to the SWAT project folder (i.e. TxtInOut)}
}
\description{
Generate folder structure for SWAT execution
}
\keyword{internal}
