% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_input_files.R
\name{set_codes_bsn}
\alias{set_codes_bsn}
\title{Set the \code{nostress} value in the codes.bsn file}
\usage{
set_codes_bsn(run_path, nostress)
}
\arguments{
\item{run_path}{Path to the folder where simulations are performed}

\item{nostress}{\code{nostress} parameter in the 'codes.bsn' file to
activate/deactivate plant stresses for plant growth.
\itemize{
\item \code{nostress = 0}: all plant stresses applied
\item \code{nostress = 1}: turn off all plant stresses
\item \code{nostress = 2}: turn off nutrient plant stress only
}}
}
\description{
Set the \code{nostress} value in the codes.bsn file
}
\keyword{internal}
