% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_input_files.R
\name{set_print_prt}
\alias{set_print_prt}
\title{Read and set SWAT+ print.prt file write the updated print.prt and the original file
as a backup 'print_backup.prt'}
\usage{
set_print_prt(project_path, run_path, outputs, years_skip)
}
\arguments{
\item{project_path}{Path to the SWAT project folder (i.e. TxtInOut)}

\item{run_path}{Path to the folder where simulations are performed}

\item{outputs}{Define the outputs that should be read after the simulation
run. The outputs that are defined here depend on the verification steps
that should be performed on the outputs.}

\item{years_skip}{(optional) Integer value to define the number of simulation
years that are skipped before writing SWAT model outputs.}
}
\description{
Read and set SWAT+ print.prt file write the updated print.prt and the original file
as a backup 'print_backup.prt'
}
\keyword{internal}
